# schema/openapi.yaml
openapi: 3.0.3
info:
  title: SurgiScribe API Documentation
  version: v1
  description: API documentation with JWT Authentication
paths:
...
  /api/v1/web/login/:
    post:
      operationId: web_login
      description: Sign in from a browser. Returns an access token in the body and
        sets the refresh token as an httpOnly, host-only cookie scoped to /api/v1/web/.
        The refresh token is never present in the body and must never be stored by
        the client.
      tags:
      - Authentication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebLoginRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/WebLoginRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/WebLoginRequestRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebLoginResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebError'
          description: Validation failed.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebError'
          description: Not authenticated.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebError'
          description: Account cannot sign in.
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebError'
          description: Rate limited.
  /api/v1/web/logout/:
    post:
      operationId: web_logout
      description: 'End the browser session: revokes the refresh token and clears
        the cookie. Idempotent — succeeds even with no cookie, an already-revoked
        token, or no access token, since the access token has often expired by the
        time a user signs out.'
      tags:
      - Authentication
      responses:
        '204':
          description: Session ended.
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebError'
          description: Rate limited.
  /api/v1/web/refresh/:
    post:
      operationId: web_token_refresh
      description: 'Exchange the refresh cookie for a new access token. The refresh
        token is rotated: the presented one is revoked and its replacement is set
        as the new cookie. The cookie is read from the request and is not a parameter.
        Any 401 clears the cookie, so a client holding a dead session reaches the
        sign-in screen instead of retrying.'
      tags:
      - Authentication
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebRefreshResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebError'
          description: Validation failed.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebError'
          description: Not authenticated.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebError'
          description: Account cannot sign in.
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebError'
          description: Rate limited.
components:
  schemas:
...
    WebError:
      type: object
      description: |-
        Stable error shape for /api/v1/web/*.

        `code` is the machine-readable discriminator to branch on; `detail` is
        display text and may change without notice.
      properties:
        code:
          type: string
          description: e.g. invalid_credentials, account_pending, refresh_token_missing,
            refresh_token_invalid, throttled.
        detail:
          type: string
        field_errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Present only for validation failures.
      required:
      - code
      - detail
    WebLoginRequestRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - email
      - password
    WebLoginResponse:
      type: object
      description: |-
        The refresh token is deliberately absent: it is set as an httpOnly
        cookie and must never be readable by scripts.
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: Bearer
        expires_in:
          type: integer
          description: Access token lifetime in seconds. Refresh before it elapses
            rather than waiting for a 401.
        user:
          $ref: '#/components/schemas/WebUser'
      required:
      - access_token
      - expires_in
      - user
    WebOrganization:
      type: object
      description: One organization membership.
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        is_primary:
          type: boolean
      required:
      - id
      - is_primary
      - name
      - role
      - type
    WebRefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: Bearer
        expires_in:
          type: integer
      required:
      - access_token
      - expires_in
    WebUser:
      type: object
      description: The signed-in user, as much as the SPA needs to render a shell.
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          nullable: true
          description: UserProfile.role — the value all API permission checks use.
        organization_name:
          type: string
          nullable: true
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/WebOrganization'
      required:
      - email
      - id
      - name
      - organization_name
      - organizations
      - role
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT