# src/api/generated/endpoints/authentication/authentication.msw.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SurgiScribe API Documentation
 * API documentation with JWT Authentication
 * OpenAPI spec version: v1
 */
import {
  faker
} from '@faker-js/faker';

import {
  HttpResponse,
  http
} from 'msw';
import type {
  RequestHandlerOptions
} from 'msw';

import type {
  WebLoginResponse,
  WebRefreshResponse
} from '../../model';


export const getWebLoginResponseMock = (overrideResponse: Partial<Extract<WebLoginResponse, object>> = {}): WebLoginResponse => ({access_token: faker.string.alpha({length: {min: 10, max: 20}}), token_type: faker.string.alpha({length: {min: 10, max: 20}}), expires_in: faker.number.int(), user: {id: faker.number.int(), email: faker.internet.email(), name: faker.string.alpha({length: {min: 10, max: 20}}), role: faker.helpers.arrayElement([faker.string.alpha({length: {min: 10, max: 20}}), null]), organization_name: faker.helpers.arrayElement([faker.string.alpha({length: {min: 10, max: 20}}), null]), organizations: Array.from({ length: faker.number.int({min: 1, max: 10}) }, (_, i) => i + 1).map(() => ({id: faker.number.int(), name: faker.string.alpha({length: {min: 10, max: 20}}), type: faker.helpers.arrayElement([faker.string.alpha({length: {min: 10, max: 20}}), null]), role: faker.helpers.arrayElement([faker.string.alpha({length: {min: 10, max: 20}}), null]), is_primary: faker.datatype.boolean()}))}, ...overrideResponse})

export const getWebTokenRefreshResponseMock = (overrideResponse: Partial<Extract<WebRefreshResponse, object>> = {}): WebRefreshResponse => ({access_token: faker.string.alpha({length: {min: 10, max: 20}}), token_type: faker.string.alpha({length: {min: 10, max: 20}}), expires_in: faker.number.int(), ...overrideResponse})


export const getWebLoginMockHandler = (overrideResponse?: WebLoginResponse | ((info: Parameters<Parameters<typeof http.post>[1]>[0]) => Promise<WebLoginResponse> | WebLoginResponse), options?: RequestHandlerOptions) => {
  return http.post('/api/v1/web/login/', async (info: Parameters<Parameters<typeof http.post>[1]>[0]) => {


    return HttpResponse.json(overrideResponse !== undefined
    ? (typeof overrideResponse === "function" ? await overrideResponse(info) : overrideResponse)
    : getWebLoginResponseMock(),
      { status: 200
      })
  }, options)
}

export const getWebLogoutMockHandler = (overrideResponse?: void | ((info: Parameters<Parameters<typeof http.post>[1]>[0]) => Promise<void> | void), options?: RequestHandlerOptions) => {
  return http.post('/api/v1/web/logout/', async (info: Parameters<Parameters<typeof http.post>[1]>[0]) => {
  if (typeof overrideResponse === 'function') {await overrideResponse(info); }

    return new HttpResponse(null,
      { status: 204
      })
  }, options)
}

export const getWebTokenRefreshMockHandler = (overrideResponse?: WebRefreshResponse | ((info: Parameters<Parameters<typeof http.post>[1]>[0]) => Promise<WebRefreshResponse> | WebRefreshResponse), options?: RequestHandlerOptions) => {
  return http.post('/api/v1/web/refresh/', async (info: Parameters<Parameters<typeof http.post>[1]>[0]) => {


    return HttpResponse.json(overrideResponse !== undefined
    ? (typeof overrideResponse === "function" ? await overrideResponse(info) : overrideResponse)
    : getWebTokenRefreshResponseMock(),
      { status: 200
      })
  }, options)
}
export const getAuthenticationMock = () => [
  getWebLoginMockHandler(),
  getWebLogoutMockHandler(),
  getWebTokenRefreshMockHandler()
]