# src/api/generated/model/webLoginRequestRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SurgiScribe API Documentation
 * API documentation with JWT Authentication
 * OpenAPI spec version: v1
 */

export interface WebLoginRequestRequest {
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  password: string;
}