# src/api/generated/model/webLoginResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SurgiScribe API Documentation
 * API documentation with JWT Authentication
 * OpenAPI spec version: v1
 */
import type { WebUser } from './webUser';

/**
 * The refresh token is deliberately absent: it is set as an httpOnly
 * cookie and must never be readable by scripts.
 */
export interface WebLoginResponse {
  access_token: string;
  token_type?: string;
  /** Access token lifetime in seconds. Refresh before it elapses rather than waiting for a 401. */
  expires_in: number;
  user: WebUser;
}