# src/api/generated/model/webRefreshResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SurgiScribe API Documentation
 * API documentation with JWT Authentication
 * OpenAPI spec version: v1
 */

export interface WebRefreshResponse {
  access_token: string;
  token_type?: string;
  expires_in: number;
}