# src/api/generated/model/webUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SurgiScribe API Documentation
 * API documentation with JWT Authentication
 * OpenAPI spec version: v1
 */
import type { WebOrganization } from './webOrganization';

/**
 * The signed-in user, as much as the SPA needs to render a shell.
 */
export interface WebUser {
  id: number;
  email: string;
  name: string;
  /**
     * UserProfile.role — the value all API permission checks use.
     * @nullable
     */
  role: string | null;
  /** @nullable */
  organization_name: string | null;
  organizations: WebOrganization[];
}